+++ packages/showcases/svelte/stories/components/indicators/VtmnBadge/VtmnBadge.stories.svelte
@@ -1,69 +1,24 @@
 <script>
   import { Meta, Template, Story } from '@storybook/addon-svelte-csf';
-
-  import { VtmnBadge } from '@vtmn/svelte';
-  import docs from './docs.mdx';
-  import {
-    parameters,
-    argTypes,
-  } from '@vtmn/showcase-core/csf/components/indicators/badge.csf';
+  import { VtmnTag } from '@vtmn/svelte';
+  import { parameters } from '@vtmn/showcase-core/csf/components/forms/text-input.csf';
 </script>
 
 <Meta
-  title="Components / Indicators / VtmnBadge"
-  component={VtmnBadge}
-  {...argTypes}
-  parameters={{
-    ...parameters,
-    docs: {
-      page: docs,
+  title="Components / Indicators / VtmnTag"
+  component={VtmnTag}
+  argTypes={{
+    slot: {
+      control: 'text',
+      description: 'Content of the slot',
+      defaultValue: 'Tag',
     },
   }}
+  {parameters}
 />
 
 <Template let:args>
-  <VtmnBadge {...args} />
+  <VtmnTag {...args}>{args.slot}</VtmnTag>
 </Template>
 
-<Story
-  name="Overview"
-  args={{
-    variant: 'default',
-    value: 6,
-  }}
-/>
-
-<Story name="Default">
-  <VtmnBadge variant="default" />
-  <VtmnBadge variant="default" value="6" />
-  <VtmnBadge variant="default" value="100" />
-</Story>
-
-<Story name="Brand">
-  <VtmnBadge variant="brand" />
-  <VtmnBadge variant="brand" value="6" />
-  <VtmnBadge variant="brand" value="100" />
-</Story>
-
-<Story
-  name="Reversed"
-  parameters={{
-    backgrounds: { default: 'primary-reversed' },
-  }}
->
-  <VtmnBadge variant="reversed" />
-  <VtmnBadge variant="reversed" value="6" />
-  <VtmnBadge variant="reversed" value="100" />
-</Story>
-
-<Story name="Accent">
-  <VtmnBadge variant="accent" />
-  <VtmnBadge variant="accent" value="6" />
-  <VtmnBadge variant="accent" value="100" />
-</Story>
-
-<Story name="Alert">
-  <VtmnBadge variant="alert" />
-  <VtmnBadge variant="alert" value="6" />
-  <VtmnBadge variant="alert" value="100" />
-</Story>
+<Story name="Overview" args={{ variant: 'accent' }} />