+++ packages/sources/assets/src/scripts/build-svg.js
@@ -1,0 +1,11 @@
+const fs = require('fs');
+const fse = require('fs-extra');
+
+const assetsSrcDir = `${__dirname}/../assets`;
+const assetsDestDir = `${__dirname}/../../dist/assets`;
+
+fse.copy(assetsSrcDir, `${assetsDestDir}/svg`, { overwrite: true }, (err) => {
+  if (err) {
+    console.error(err);
+  }
+});