+++ packages/sources/svelte/rollup.config.js
@@ -1,6 +1,6 @@
-import svelte from 'rollup-plugin-svelte';
 import resolve from '@rollup/plugin-node-resolve';
 import css from 'rollup-plugin-css-only';
+import svelte from 'rollup-plugin-svelte';
 import sveltePreprocess from 'svelte-preprocess';
 const preprocessOptions = require('./svelte.config').preprocessOptions;
 
@@ -18,8 +18,8 @@ const components = [
     name: 'VtmnTextInput',
   },
   {
-    folder: 'overlays',
-    name: 'VtmnPopover',
+    folder: 'indicators',
+    name: 'VtmnBadge',
   },
   {
     folder: 'indicators',
@@ -27,7 +27,23 @@ const components = [
   },
   {
     folder: 'indicators',
-    name: 'VtmnBadge',
+    name: 'VtmnTag',
+  },
+  {
+    folder: 'overlays',
+    name: 'VtmnPopover',
+  },
+  {
+    folder: 'selection-controls',
+    name: 'VtmnCheckbox',
+  },
+  {
+    folder: 'structure',
+    name: 'VtmnList',
+  },
+  {
+    folder: 'structure',
+    name: 'VtmnListItem',
   },
 ];