+++ packages/sources/svelte/src/components/index.js
@@ -8,6 +8,7 @@ export { default as VtmnTextInput } from './forms/VtmnTextInput/VtmnTextInput.sv
 // Indicators
 export { default as VtmnBadge } from './indicators/VtmnBadge/VtmnBadge.svelte';
 export { default as VtmnPrice } from './indicators/VtmnPrice/VtmnPrice.svelte';
+export { default as VtmnTag } from './indicators/VtmnTag/VtmnTag.svelte';
 
 // Navigation
 
@@ -15,5 +16,8 @@ export { default as VtmnPrice } from './indicators/VtmnPrice/VtmnPrice.svelte';
 export { default as VtmnPopover } from './overlays/VtmnPopover/VtmnPopover.svelte';
 
 // Selection controls
+export { default as VtmnCheckbox } from './selection-controls/VtmnCheckbox/VtmnCheckbox.svelte';
 
 // Structure
+export { default as VtmnList } from './structure/VtmnList/VtmnList.svelte';
+export { default as VtmnListItem } from './structure/VtmnListItem/VtmnListItem.svelte';