+++ packages/sources/vue/src/components/index.ts
@@ -2,6 +2,9 @@ import VtmnButton from './actions/VtmnButton';
 import VtmnLink from './actions/VtmnLink';
 import VtmnTextInput from './forms/VtmnTextInput';
 import VtmnSelect from './forms/VtmnSelect';
+import VtmnBadge from './indicators/VtmnBadge';
+import VtmnLoader from './indicators/VtmnLoader';
+import VtmnPrice from './indicators/VtmnPrice';
 import VtmnCheckbox from './selection-controls/VtmnCheckbox';
 import VtmnRadioButton from './selection-controls/VtmnRadioButton';
 import VtmnToggle from './selection-controls/VtmnToggle';
@@ -11,6 +14,9 @@ export {
   VtmnLink,
   VtmnTextInput,
   VtmnSelect,
+  VtmnBadge,
+  VtmnLoader,
+  VtmnPrice,
   VtmnCheckbox,
   VtmnRadioButton,
   VtmnToggle,