+++ packages/sources/vue/src/components/indicators/VtmnBadge/types.ts
@@ -1,0 +1,6 @@
+export type VtmnBadgeVariant =
+  | 'default'
+  | 'brand'
+  | 'reversed'
+  | 'accent'
+  | 'alert';