+++ packages/sources/vue/src/index.ts
@@ -28,6 +28,9 @@ export { default as VtmnSelect } from './components/forms/VtmnSelect';
 export { default as VtmnTextInput } from './components/forms/VtmnTextInput';
 
 // Indicators
+export { default as VtmnBadge } from './components/indicators/VtmnBadge';
+export { default as VtmnLoader } from './components/indicators/VtmnLoader';
+export { default as VtmnPrice } from './components/indicators/VtmnPrice';
 
 // Navigation