+++ packages/sources/css/src/components/navigation/breadcrumb/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.13](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-breadcrumb@0.3.12...@vtmn/css-breadcrumb@0.3.13) (2022-03-01)
+
+**Note:** Version bump only for package @vtmn/css-breadcrumb
+
+
+
+
+
 ## [0.3.12](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-breadcrumb@0.3.11...@vtmn/css-breadcrumb@0.3.12) (2022-03-01)
 
 **Note:** Version bump only for package @vtmn/css-breadcrumb