+++ packages/sources/css/src/components/navigation/navbar/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.16](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-navbar@0.1.15...@vtmn/css-navbar@0.1.16) (2022-03-01)
+
+**Note:** Version bump only for package @vtmn/css-navbar
+
+
+
+
+
 ## [0.1.15](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-navbar@0.1.14...@vtmn/css-navbar@0.1.15) (2022-03-01)
 
 **Note:** Version bump only for package @vtmn/css-navbar