+++ packages/sources/css/src/components/overlays/toast/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.9](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-toast@0.5.8...@vtmn/css-toast@0.5.9) (2022-03-01)
+
+**Note:** Version bump only for package @vtmn/css-toast
+
+
+
+
+
 ## [0.5.8](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-toast@0.5.7...@vtmn/css-toast@0.5.8) (2022-03-01)
 
 **Note:** Version bump only for package @vtmn/css-toast