+++ packages/sources/css/src/components/selection-controls/checkbox/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.13](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-checkbox@0.6.12...@vtmn/css-checkbox@0.6.13) (2022-03-01)
+
+**Note:** Version bump only for package @vtmn/css-checkbox
+
+
+
+
+
 ## [0.6.12](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-checkbox@0.6.11...@vtmn/css-checkbox@0.6.12) (2022-03-01)
 
 **Note:** Version bump only for package @vtmn/css-checkbox