+++ packages/sources/vue/CHANGELOG.md
@@ -3,6 +3,120 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.18.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.17.0...@vtmn/vue@0.18.0) (2022-03-03)
+
+
+### Features
+
+* **@vtmn/vue:** add `VtmnBadge` component ([#1007](https://github.com/Decathlon/vitamin-web/issues/1007)) ([b535717](https://github.com/Decathlon/vitamin-web/commit/b535717c4584d5649d4fa600f538fd8ce7ac47ca))
+
+
+
+
+
+# [0.17.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.16.1...@vtmn/vue@0.17.0) (2022-03-03)
+
+
+### Features
+
+* **@vtmn/vue:** add `VtmnPrice` component ([#1006](https://github.com/Decathlon/vitamin-web/issues/1006)) ([99b0136](https://github.com/Decathlon/vitamin-web/commit/99b01360db6057950a7e22b032e34931fa81c1aa))
+
+
+
+
+
+## [0.16.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.16.0...@vtmn/vue@0.16.1) (2022-03-02)
+
+**Note:** Version bump only for package @vtmn/vue
+
+
+
+
+
+# [0.16.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.15.4...@vtmn/vue@0.16.0) (2022-03-02)
+
+
+### Features
+
+* **@vtmn/vue:** display icon with `VtmnIcon`  and remove useless type in `VtmnButton` ([#999](https://github.com/Decathlon/vitamin-web/issues/999)) ([b9ab259](https://github.com/Decathlon/vitamin-web/commit/b9ab259c1a6914d61aeeb1a0e394cd6312d15e62))
+
+
+
+
+
+## [0.15.4](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.15.3...@vtmn/vue@0.15.4) (2022-03-02)
+
+**Note:** Version bump only for package @vtmn/vue
+
+
+
+
+
+## [0.15.3](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.15.2...@vtmn/vue@0.15.3) (2022-03-02)
+
+**Note:** Version bump only for package @vtmn/vue
+
+
+
+
+
+## [0.15.2](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.15.1...@vtmn/vue@0.15.2) (2022-03-01)
+
+**Note:** Version bump only for package @vtmn/vue
+
+
+
+
+
+## [0.15.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.15.0...@vtmn/vue@0.15.1) (2022-03-01)
+
+**Note:** Version bump only for package @vtmn/vue
+
+
+
+
+
+# [0.15.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.14.4...@vtmn/vue@0.15.0) (2022-03-01)
+
+
+### Features
+
+* **@vtmn/vue:** add `VtmnLoader` component ([#1003](https://github.com/Decathlon/vitamin-web/issues/1003)) ([2a05e77](https://github.com/Decathlon/vitamin-web/commit/2a05e776090472bf0b33c933e78cfe97bbae2c4b))
+
+
+
+
+
+## [0.14.4](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.14.3...@vtmn/vue@0.14.4) (2022-03-01)
+
+**Note:** Version bump only for package @vtmn/vue
+
+
+
+
+
+## [0.14.3](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.14.2...@vtmn/vue@0.14.3) (2022-03-01)
+
+
+### Bug Fixes
+
+* **@vtmn/vue:** sonarcloud duplications ([c973e05](https://github.com/Decathlon/vitamin-web/commit/c973e0506397c2b8684012f785d87c1c50242ba2))
+
+
+
+
+
+## [0.14.2](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.14.1...@vtmn/vue@0.14.2) (2022-03-01)
+
+
+### Bug Fixes
+
+* **@vtmn/vue:** sonarcloud duplications for forms components ([ea9a17e](https://github.com/Decathlon/vitamin-web/commit/ea9a17edd0cefd90b5faf067f6af94b06c3ff318))
+
+
+
+
+
 ## [0.14.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.14.0...@vtmn/vue@0.14.1) (2022-03-01)
 
 **Note:** Version bump only for package @vtmn/vue